try to fix windows ci issue

+++ cpp/open3d/t/io/NumpyIO.cpp
@@ -289,7 +289,7 @@ ParseNpyHeaderFromBuffer(const char* buffer) {
 static std::tuple<uint16_t, size_t, size_t> ParseZipFooter(FILE* fp) {
     size_t footer_len = 22;
     std::vector<char> footer(footer_len);
-    fseek(fp, -footer_len, SEEK_END);
+    fseek(fp, -static_cast<int64_t>(footer_len), SEEK_END);
     if (fread(footer.data(), sizeof(char), footer_len, fp) != footer_len) {
         utility::LogError("Footer fread failed.");
     }
